Create hardlinks only if missing

+++ developer-mode/builds/build-tcl.sh
@@ -22,9 +22,13 @@ echo "Building TCL"
 # cut down on the output to stdout to make Travis-CI consoles faster
 make 2>&1 | tee -a /workspaces/logs/tcl.log | cut -c1-64
 make install
-# Link to tclsh
-ln /usr/bin/tclsh8.6 /bin/tclsh
-ln /usr/bin/tclsh8.6 /usr/bin/tclsh
+# Link to tclsh if needed
+if [ ! -f /bin/tclsh ]; then
+    ln /usr/bin/tclsh8.6 /bin/tclsh
+fi
+if [ ! -f /usr/bin/tclsh ]; then
+    ln /usr/bin/tclsh8.6 /usr/bin/tclsh
+fi 
 
 build_cleanup
